Moving CraftItem into CraftingLogic
